Refactor recommendation hooks
